backend/requirements.txt
flask>=2.0.0
flask-cors>=3.0.10
pymongo>=4.0.0
python-dotenv>=0.19.0
pytest>=7.0.0
robotframework>=6.0.0
robotframework-seleniumlibrary>=6.0.0
beautifulsoup4>=4.9.3
bleach>=4.1.0

